requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
motor==3.7.0
langchain==0.3.18
langchain_community==0.3.17
transformers==4.48.3
faiss-cpu==1.10.0
python-dotenv==1.0.1
pymongo==4.11.1
python-multipart==0.0.20
sentence-transformers==3.4.1
openai==1.63.0
pydantic==2.10.6
websockets==15.0
mutagen==1.47.0
azure-cognitiveservices-speech==1.42.0
starlette==0.45.3
transitions==0.9.2
langdetect==1.0.9
soundfile==0.13.1
sounddevice==0.5.1
speechrecognition==3.14.1
pydantic-settings
huggingface-hub
pypdf
numpy
torch
Pillow
pymupdf
psutil
httpx
requests
azure-storage-blob
langchain-huggingface
googletrans